Revert the conversion from u32 to size_t from my last commit to reduce compiler warnings in windows. Fix an issue that made it impossible to close the console and log windows in windows.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@5989 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DolphinWX/Src/Frame.h
@@ -145,7 +145,7 @@ class CFrame : public CRenderFrame
 
 		// Utility
 		wxWindow * GetNotebookPageFromId(wxWindowID Id);
-		wxAuiNotebook * GetNotebookFromId(size_t NBId);
+		wxAuiNotebook * GetNotebookFromId(u32 NBId);
 		int GetNotebookCount();
 		wxString GetMenuLabel(int Id);
